chore: added .pre-commit for remove white space
