+++ .idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,43 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="16">
+            <item index="0" class="java.lang.String" itemvalue="importlib_metadata" />
+            <item index="1" class="java.lang.String" itemvalue="jupyterlab_pygments" />
+            <item index="2" class="java.lang.String" itemvalue="importlib_resources" />
+            <item index="3" class="java.lang.String" itemvalue="jupyter_client" />
+            <item index="4" class="java.lang.String" itemvalue="typing_extensions" />
+            <item index="5" class="java.lang.String" itemvalue="ipython" />
+            <item index="6" class="java.lang.String" itemvalue="jupyter_core" />
+            <item index="7" class="java.lang.String" itemvalue="fonttools" />
+            <item index="8" class="java.lang.String" itemvalue="zipp" />
+            <item index="9" class="java.lang.String" itemvalue="prompt-toolkit" />
+            <item index="10" class="java.lang.String" itemvalue="requests" />
+            <item index="11" class="java.lang.String" itemvalue="fsspec" />
+            <item index="12" class="java.lang.String" itemvalue="tornado" />
+            <item index="13" class="java.lang.String" itemvalue="certifi" />
+            <item index="14" class="java.lang.String" itemvalue="packaging" />
+            <item index="15" class="java.lang.String" itemvalue="urllib3-secure-extra" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
+    <inspection_tool class="PyStubPackagesAdvertiser" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <list>
+          <option value="pyspark-stubs==3.0.0.post3" />
+        </list>
+      </option>
+    </inspection_tool>
+    <inspection_tool class="PyUnresolvedReferencesInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredIdentifiers">
+        <list>
+          <option value="str.*" />
+        </list>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
