+++ abacatepay/domain/entities/product.py
@@ -1,0 +1,10 @@
+from typing import List, Union
+from pydantic import BaseModel
+
+
+class Product(BaseModel):
+    externalId: str
+    name: str
+    quantity: int
+    price: int
+    description: str
