+++ abacatepay/domain/response_models.py
@@ -1,48 +1,6 @@
 from typing import List, Union
-from ._constants import BILLING_STATUS, BILLING_METHODS, BILLING_KINDS
-from pydantic import BaseModel
-
-
-class Product(BaseModel):
-    externalId: str
-    name: str
-    quantity: int
-    price: int
-    description: str
-
-
-class Customer(BaseModel):
-    """
-    Customer model
-
-    Attributes:
-        id: str | None
-        taxId: str | None
-        name: str | None
-        email: str
-        cellphone: str | None
-    Note: never pass `id` as an argument, it is supposed to be returned by the API!
-    """
-    id: str | None = None
-    taxId: str | None = None
-    name: str | None = None
-    email: str
-    cellphone: str | None = None
-
-    @classmethod
-    def from_dict(cls, data: dict):
-        metadata = data.get("metadata", {})
-        return cls(
-            id = data.get("id"),
-            taxId=metadata.get("taxId"),
-            name=metadata.get("name"),
-            email=metadata.get("email"),
-            cellphone=metadata.get("cellphone"),
-        )
-
-
-
-
+from abacatepay.domain.entities.customer import Customer
+from abacatepay.infrastructure.utils.constants import BILLING_STATUS, BILLING_METHODS, BILLING_KINDS
 
 class BillingResponse:
     def __init__(self, data: dict):
